changing id from int to Long

+++ src/main/java/com/utn/appparcial/model/Professor.java
@@ -5,7 +5,7 @@ public class Professor extends User {
     public Professor() {
     }
 
-    public Professor(int id, String name, String lastName, String email, String password) {
+    public Professor(Long id, String name, String lastName, String email, String password) {
         super(id, name, lastName, email, password);
     }
